backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
pymongo==4.7.3
PyJWT==2.8.0
bcrypt==4.1.3
python-dotenv==1.0.1
PyPDF2==3.0.1
scikit-learn==1.5.0
nltk==3.8.1
Flask-Mail==0.10.0
